+++ server/src/events/entity_manipulation.rs
@@ -211,7 +211,7 @@ pub fn handle_explosion(server: &Server, pos: Vec3<f32>, power: f32, owner: Opti
     for (pos_b, ori_b, character_b, stats_b) in (
         &ecs.read_storage::<comp::Pos>(),
         &ecs.read_storage::<comp::Ori>(),
-        &ecs.read_storage::<comp::CharacterState>(),
+        ecs.read_storage::<comp::CharacterState>().maybe(),
         &mut ecs.write_storage::<comp::Stats>(),
     )
         .join()
@@ -231,7 +231,7 @@ pub fn handle_explosion(server: &Server, pos: Vec3<f32>, power: f32, owner: Opti
             }
 
             // Block
-            if character_b.is_block()
+            if character_b.map(|c_b| c_b.is_block()).unwrap_or(false)
                 && ori_b.0.angle_between(pos - pos_b.0) < BLOCK_ANGLE.to_radians() / 2.0
             {
                 dmg = (dmg as f32 * (1.0 - BLOCK_EFFICIENCY)) as u32
@@ -261,7 +261,7 @@ pub fn handle_explosion(server: &Server, pos: Vec3<f32>, power: f32, owner: Opti
             .read_resource::<TerrainGrid>()
             .ray(pos, pos + dir * color_range)
             .until(|_| rand::random::<f32>() < 0.05)
-            .for_each(|pos| touched_blocks.push(pos))
+            .for_each(|_: &Block, pos| touched_blocks.push(pos))
             .cast();
     }
 
@@ -296,12 +296,8 @@ pub fn handle_explosion(server: &Server, pos: Vec3<f32>, power: f32, owner: Opti
         let _ = terrain
             .ray(pos, pos + dir * power)
             .until(|block| block.is_fluid() || rand::random::<f32>() < 0.05)
-            .for_each(|pos| {
-                if terrain
-                    .get(pos)
-                    .map(|block| block.is_explodable())
-                    .unwrap_or(false)
-                {
+            .for_each(|block: &Block, pos| {
+                if block.is_explodable() {
                     block_change.set(pos, Block::empty());
                 }
             })