+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new animals
 - Better pathfinding
 - Bombs
+- Training dummies
 
 ### Changed